sorakun_apiTest create common post and get API
